Split source files to support unit tests
