+++ src/secplus2.h
@@ -1,0 +1,7 @@
+#ifndef _SECPLUS2_H
+#define _SECPLUS2_H
+
+const uint8_t SECPLUS2_CODE_LEN = 19;
+const uint32_t SECPLUS2_PREAMBLE = 0x00550100;
+
+#endif // _SECPLUS2_H
